decomposed benchmark.py
